# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mobile App API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/set-order": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Set an order",
                "operationId": "230fa2c0c1ce65e2a71a5db180f56b34",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "service_id": {
                                        "type": "integer"
                                    },
                                    "reserve_datetime": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "address_id": {
                                        "type": "integer"
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "payment_method": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order set successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "409": {
                        "description": "Order already exists"
                    }
                }
            }
        },
        "/api/cancel-order": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cancel an order",
                "operationId": "dddb8f0f4351311ca1d909d43fe99064",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order cancelled successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/api/calculate-service-fees": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Calculate service fees",
                "operationId": "fb8e4655d7299931447a03e31d9d42e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "service_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service fees calculated successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Service not found"
                    }
                }
            }
        },
        "/api/verify-coupon": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Verify a coupon",
                "operationId": "68d29187a2414b2d7241fee5f0a8806e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "coupon": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Coupon verified successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Coupon not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}